requirements.txt
annotated-types==0.6.0
contourpy==1.2.0
cycler==0.12.1
fonttools==4.47.2
iperf3==0.1.11
kiwisolver==1.4.5
matplotlib==3.8.2
numpy==1.26.3
packaging==23.2
pandas==2.2.0
pillow==10.2.0
pydantic==1.10.0
pydantic_core==2.16.1
pyparsing==3.1.1
PyQt5==5.15.10
PyQt5-Qt5==5.15.2
PyQt5-sip==12.13.0
python-dateutil==2.8.2
pytz==2023.4
redbird==0.7.1
rocketry==2.5.1
six==1.16.0
speedtest-cli==2.1.3
typing_extensions==4.9.0
tzdata==2023.4
